requirements.txt
amqp==5.1.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
billiard==4.1.0
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cloudinary==1.35.0
decorator==4.4.2
dnspython==2.4.2
docutils==0.19
exceptiongroup==1.1.3
extras==1.0.0
fastapi==0.103.1
future==0.18.3
greenlet==2.0.2
h11==0.14.0
idna==3.4
imageio==2.31.4
imageio-ffmpeg==0.4.9
kombu==5.3.2
lockfile==0.12.2
moviepy==1.0.3
numpy==1.26.0
pbr==5.9.0
Pillow==10.0.1
proglog==0.1.10
prompt-toolkit==3.0.39
protobuf==4.21.12
psycopg2-binary==2.9.8
pydantic==2.4.2
pydantic_core==2.10.1
pymongo==4.5.0
python-daemon==2.3.1
python-dateutil==2.8.2
python-decouple==3.8
python-ffmpeg-video-streaming==0.1.15
python-multipart==0.0.6
requests==2.31.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.21
starlette==0.27.0
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.16
uvicorn==0.23.2
vine==5.0.0
wcwidth==0.2.8
